+++ src/test/java/de/caritas/cob/consultingtypeservice/api/mapper/ExtendedConsultingTypeMapperTest.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 public class ExtendedConsultingTypeMapperTest {
 
   @Test
-  public void mapConsultingType_Should_MapAllBasicProperties() throws IOException {
+  public void mapConsultingType_Should_MapAllExtendedProperties() throws IOException {
 
     ExtendedConsultingTypeResponseDTO result = ExtendedConsultingTypeMapper
         .mapConsultingType(HelperMethods.getConsultingType());
@@ -24,7 +24,8 @@ public class ExtendedConsultingTypeMapperTest {
         jsonEquals(HelperMethods.getConsultingTypeSettingsAsJsonString())
             .whenIgnoringPaths(
                 "voluntaryComponents",
-                "requiredComponents"
+                "requiredComponents",
+                "welcomeScreen"
             )
     );
   }